+++ src/main/java/kyrie/ByeCommand.java
@@ -1,3 +1,5 @@
+package kyrie;
+
 public class ByeCommand extends Command {
     public ByeCommand() {
         super();
