+++ src/main/java/kyrie/DeleteCommand.java
@@ -1,3 +1,5 @@
+package kyrie;
+
 public class DeleteCommand extends Command {
     private int index;
