+++ src/main/java/kyrie/Event.java
@@ -1,3 +1,5 @@
+package kyrie;
+
 public class Event extends Task {
     protected DateTime start;
     protected DateTime end;
