+++ src/main/java/kyrie/Kyrie.java
@@ -1,3 +1,5 @@
+package kyrie;
+
 import java.io.File;
 import java.io.IOException;
 import java.util.Scanner;
