+++ src/main/java/kyrie/MarkCommand.java
@@ -1,15 +1,17 @@
-public class UnmarkCommand extends Command {
+package kyrie;
+
+public class MarkCommand extends Command {
     private Integer index;
 
-    public UnmarkCommand(Integer index) {
+    public MarkCommand(Integer index) {
         super();
         this.index = index;
     }
 
     @Override
     public void execute(Storage storage, Ui ui, TaskList tasks) {
-        tasks.unmarkTaskAsDone(index);
-        ui.showMarkedTask(tasks.getTask(index));
+        tasks.markTaskAsDone(this.index);
+        ui.showMarkedTask(tasks.getTask(this.index));
         storage.saveData(tasks);
     }
 }
