+++ src/main/java/kyrie/TaskType.java
@@ -1,3 +1,5 @@
+package kyrie;
+
 public enum TaskType {
     TODO,
     DEADLINE,
