+++ src/main/java/kyrie/Ui.java
@@ -1,3 +1,5 @@
+package kyrie;
+
 import java.util.Scanner;
 
 public class Ui {
